fix course edit v2

+++ app/Http/Controllers/CourseController.php
@@ -162,13 +162,23 @@ protected $result=[];
             'picture'=>$request->picture,
 
         ];
-        if($data['type']=='course'){
-          $product= Product::query()->where('course_id',$id)->firstOrFail()->toArray();
-        $pro_id=$product['id'];
-        Invoice::query()->where('order_id',$pro_id)->firstOrFail()->delete();
-        DB::table('bookmarks')->where('bookmarkable_id',$pro_id)->delete();
-        Product::query()->where('course_id',$id)->firstOrFail()->delete();
+       if($data['type']=='course') {
+        $product=Product::query()->where('course_id',$id)->firstOrFail()->toArray();
+        if($product){
+            $pro_id=$product['id'];
+            Invoice::query()->whereIn('order_id',$pro_id)->delete();
+            DB::table('bookmarks')->where('bookmarkable_id',$pro_id)->delete();
+            Product::query()->where('course_id',$id)->delete();
         }
+       }
+
+//        if($data['type']=='course'){
+//           $product= Product::query()->where('course_id',$id)->first()->toArray();
+//        $pro_id=$product['id'];
+//        Invoice::query()->where('order_id',$pro_id)->delete();
+//        DB::table('bookmarks')->where('bookmarkable_id',$pro_id)->delete();
+//        Product::query()->where('course_id',$id)->delete();
+//        }
 
         $categories=explode(",",$request->categories);
         $Course = $this->courseRepository->update($id,$data);
